Bug fixed

+++ handler/down.go
@@ -14,6 +14,7 @@ type (
 	DownloadError struct {
 		FileName string `json:"filename"`
 		URL      string `json:"url,omitempty"`
+		Code     int    `json:"code"`
 		Reason   string `json:"reason"`
 	}
 )
@@ -21,23 +22,30 @@ type (
 func SingleDownload(mp3List []*provider.MP3) {
 	total, success, failure, ignore := len(mp3List), 0, 0, 0
 
-	var failureInfo []*DownloadError
+	dlErrs := make([]*DownloadError, 0)
 	for _, m := range mp3List {
 		switch status := m.SingleDownload(); status {
 		case ecode.Success:
 			success++
-		case ecode.AlreadyDownloaded, ecode.SongUnavailable:
+		case ecode.AlreadyDownloaded:
 			ignore++
 		default:
 			failure++
-			failureInfo = append(failureInfo, &DownloadError{m.FileName, m.DownloadURL, ecode.Message(status)})
-			// ignore error
-			os.Remove(filepath.Join(m.SavePath, m.FileName))
+			dlErrs = append(dlErrs, &DownloadError{
+				FileName: m.FileName,
+				URL:      m.DownloadURL,
+				Code:     status,
+				Reason:   ecode.Message(status),
+			})
+			if status != ecode.SongUnavailable {
+				// ignore error
+				os.Remove(filepath.Join(m.SavePath, m.FileName))
+			}
 		}
 	}
 
 	fmt.Printf("\nDownload report --> total: %d, success: %d, failure: %d, ignore: %d\n", total, success, failure, ignore)
-	outputLog(failureInfo)
+	outputLog(dlErrs)
 }
 
 func ConcurrentDownload(mp3List []*provider.MP3, n int) {
@@ -51,22 +59,29 @@ func ConcurrentDownload(mp3List []*provider.MP3, n int) {
 	}
 	c.Wait()
 
-	var failureInfo []*DownloadError
+	dlErrs := make([]*DownloadError, 0)
 	for range mp3List {
 		task := <-taskList
 		switch task.Status {
 		case ecode.Success:
 			success++
-		case ecode.AlreadyDownloaded, ecode.SongUnavailable:
+		case ecode.AlreadyDownloaded:
 			ignore++
 		default:
 			failure++
-			failureInfo = append(failureInfo, &DownloadError{task.MP3.FileName, task.MP3.DownloadURL, ecode.Message(task.Status)})
-			// ignore error
-			os.Remove(filepath.Join(task.MP3.SavePath, task.MP3.FileName))
+			dlErrs = append(dlErrs, &DownloadError{
+				FileName: task.MP3.FileName,
+				URL:      task.MP3.DownloadURL,
+				Code:     task.Status,
+				Reason:   ecode.Message(task.Status),
+			})
+			if task.Status != ecode.SongUnavailable {
+				// ignore error
+				os.Remove(filepath.Join(task.MP3.SavePath, task.MP3.FileName))
+			}
 		}
 	}
 
 	fmt.Printf("\nDownload report --> total: %d, success: %d, failure: %d, ignore: %d\n", total, success, failure, ignore)
-	outputLog(failureInfo)
+	outputLog(dlErrs)
 }
